requirements.txt
aws-cdk-lib==2.198.0
constructs>=10.0.0,<11.0.0
boto3
pandas
bs4
requests

